Finish part 4d.

+++ part4/bloglist/controllers/blogs.js
@@ -1,26 +1,40 @@
 import express from 'express'
+import jwt from 'jsonwebtoken'
+
 import Blog from '../models/blog.js'
+import User from '../models/user.js'
 
 const blogsRouter = express.Router()
 
 blogsRouter.get('/', async (req, res) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     res.json(blogs)
 })
 
 blogsRouter.post('/', async (req, res, next) => {
     const body = req.body
 
-    const blog = new Blog({
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes || 0,
-    })
-
     try {
-        const result = await blog.save()
-        res.status(201).json(result)
+        const decodedToken = jwt.verify(req.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return res.status(401).json({ error: 'token missing or invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+
+        const blog = new Blog({
+            title: body.title,
+            author: body.author,
+            url: body.url,
+            likes: body.likes || 0,
+            user: user._id,
+        })
+
+        const savedBlog = await blog.save()
+
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+
+        res.status(201).json(savedBlog)
     } catch (exception) {
         next(exception)
     }
@@ -28,8 +42,20 @@ blogsRouter.post('/', async (req, res, next) => {
 
 blogsRouter.delete('/:id', async (req, res, next) => {
     try {
-        await Blog.findByIdAndRemove(req.params.id)
-        res.status(204).end()
+        const decodedToken = jwt.verify(req.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return res.status(401).json({ error: 'token missing or invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(req.params.id)
+
+        if (blog.user.toString() === user._id.toString()) {
+            await Blog.findByIdAndRemove(req.params.id)
+            return res.status(204).end()
+        } else {
+            return res.status(401).json({ error: 'cannot delete - wrong user' })
+        }
+
     } catch (exception) {
         next(exception)
     }
